feat(Bidding): Bidding Page
