+++ components/Page.tsx
@@ -17,6 +17,10 @@ const Page: React.FC<TProps> = ({ title, children }) => {
           href="https://fonts.googleapis.com/css?family=Roboto&display=swap"
           rel="stylesheet"
         />
+        <link
+          rel="stylesheet"
+          href="https://fonts.googleapis.com/icon?family=Material+Icons"
+        />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
       </Head>
       <div className="container">
